+++ Plugins/ChromaSDKPlugin/Source/ChromaSDKPlugin/Private/Animation2D.cpp
@@ -178,7 +178,7 @@ void Animation2D::Update(float deltaTime)
 	if (_mCurrentFrame == -1)
 	{
 		_mCurrentFrame = 0;
-		if (_mCurrentFrame < _mEffects.size())
+		if (_mCurrentFrame < (int)_mEffects.size())
 		{
 			FChromaSDKEffectResult& effect = _mEffects[_mCurrentFrame];
 			int result = UChromaSDKPluginBPLibrary::ChromaSDKSetEffect(effect.EffectId);
@@ -196,7 +196,7 @@ void Animation2D::Update(float deltaTime)
 		{
 			_mTime = 0.0f;
 			++_mCurrentFrame;
-			if (_mCurrentFrame < _mEffects.size())
+			if (_mCurrentFrame < (int)_mEffects.size())
 			{
 				FChromaSDKEffectResult& effect = _mEffects[_mCurrentFrame];
 				int result = UChromaSDKPluginBPLibrary::ChromaSDKSetEffect(effect.EffectId);
